# packages/api-client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Shapers.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/otp/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "RequestSignInCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCodeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/otp/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "VerifySignInCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "RefreshSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/auth/staff/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "StaffLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/staff/set-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "SetStaffPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/auth/2fa/enable": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "EnableTwoFactor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableTwoFactorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/preferences": {
      "put": {
        "tags": [
          "My account"
        ],
        "operationId": "SetMyPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/admin/roles": {
      "get": {
        "tags": [
          "Access"
        ],
        "operationId": "ListRoles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Access"
        ],
        "operationId": "CreateRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/roles/{id}": {
      "put": {
        "tags": [
          "Access"
        ],
        "operationId": "UpdateRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/people/{personId}/access": {
      "get": {
        "tags": [
          "Access"
        ],
        "operationId": "GetPersonAccess",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonAccessDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/people/{personId}/staff-login": {
      "post": {
        "tags": [
          "Access"
        ],
        "operationId": "PrepareStaffLogin",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffLoginSetupDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grants": {
      "post": {
        "tags": [
          "Access"
        ],
        "operationId": "CreateGrant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/grants/{id}": {
      "delete": {
        "tags": [
          "Access"
        ],
        "operationId": "RevokeGrant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConsentDecision": {
        "required": [
          "purpose",
          "granted"
        ],
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string"
          },
          "granted": {
            "type": "boolean"
          }
        }
      },
...
      "CreateGrantRequest": {
        "required": [
          "personId",
          "roleId",
          "scope",
          "expiresAt",
          "reason"
        ],
        "type": "object",
        "properties": {
          "personId": {
            "type": "string",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "string"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EnableTwoFactorRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
      "GrantDto": {
        "required": [
          "id",
          "roleId",
          "roleName",
          "scope",
          "grantedAt",
          "expiresAt",
          "reason",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "roleName": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "grantedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "LogoutRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
      "PersonAccessDto": {
        "required": [
          "personId",
          "userId",
          "email",
          "phone",
          "twoFactorEnabled",
          "hasPassword",
          "lastSignInAt",
          "grants"
        ],
        "type": "object",
        "properties": {
          "personId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "lastSignInAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantDto"
            }
          }
        }
      },
...
      "PreferencesRequest": {
        "required": [
          "palette"
        ],
        "type": "object",
        "properties": {
          "palette": {
            "type": "string"
          }
        }
      },
...
      "RecoveryCodes": {
        "required": [
          "codes"
        ],
        "type": "object",
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RefreshRequest": {
        "required": [
          "refreshToken",
          "device"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          },
          "device": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "challengeId",
          "registrationTicket",
          "firstName",
          "lastName",
          "email",
          "campusId",
          "policyVersion",
          "consents",
          "device"
        ],
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "registrationTicket": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "campusId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "policyVersion": {
            "type": "string"
          },
          "consents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsentDecision"
            }
          },
          "device": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RequestCodeRequest": {
        "required": [
          "phone"
        ],
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          }
        }
      },
      "RequestCodeResponse": {
        "required": [
          "challengeId",
          "expiresAt",
          "maskedPhone"
        ],
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "maskedPhone": {
            "type": "string"
          }
        }
      },
      "RoleDto": {
        "required": [
          "id",
          "name",
          "description",
          "isSystem",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isSystem": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SaveRoleRequest": {
        "required": [
          "name",
          "description",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SetPasswordRequest": {
        "required": [
          "userId",
          "token",
          "password"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SignInResponse": {
        "required": [
          "status",
          "tokens",
          "registrationTicket",
          "existingRecordFound"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SignInStatus"
          },
          "tokens": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TokenPair"
              }
            ]
          },
          "registrationTicket": {
            "type": [
              "null",
              "string"
            ]
          },
          "existingRecordFound": {
            "type": "boolean"
          }
        }
      },
      "SignInStatus": {
        "enum": [
          "SignedIn",
          "RegistrationRequired"
        ]
      },
      "StaffLoginRequestDto": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "StaffLoginResponse": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StaffLoginStatus"
          }
        }
      },
      "StaffLoginSetupDto": {
        "required": [
          "userId",
          "email",
          "setupToken"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "setupToken": {
            "type": "string"
          }
        }
      },
      "StaffLoginStatus": {
        "enum": [
          "SignedIn",
          "TwoFactorRequired"
        ]
      },
      "StaffSignInRequest": {
        "required": [
          "email",
          "password",
          "twoFactorCode",
          "recoveryCode"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "twoFactorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "recoveryCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TokenPair": {
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "VerifyCodeRequest": {
        "required": [
          "challengeId",
          "code",
          "device"
        ],
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "device": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}